build: don't relaunch if not resetting
If we're not resetting, we can save significant run time by not running
`tutor local launch`, instead doing a `tutor local stop` followed by
`tutor local start -d`.

(This is largely because a `launch` will trigger the mfe plugin to
rebuild everything, including the -dev images, which we don't need
here.)

Also sync the PRUNE input default with the in-place one.